# openapi.yml
openapi: "3.0.0"
info:
  version: v0
  title: youyuan core Service
  license:
    name: MIT
servers:
  - url: https://api.youyuan.36node.com{basePath}
    description: The production API server
    variables:
      basePath:
        default: /core/v0
paths:
...
  /vehicles:
    post:
      summary: Create a vehicle
      operationId: createVehicle
      tags:
        - vehicle
      requestBody:
        description: vehicle to add
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleCreateDoc"
      responses:
        "201":
          description: The Vehicle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List all vehicles
      operationId: listVehicles
      tags:
        - vehicle
      parameters:
        - name: _limit
          in: query
          description: How many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            maximum: 1000
        - name: _offset
          in: query
          description: How many items to escape
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: _sort
          in: query
          description: Fields to sort
          required: false
          schema:
            type: string
        - name: _select
          in: query
          description: Fields to select
          required: false
          schema:
            type: array
            items:
              type: string
        - name: ticket
          in: query
          description: vehicle ticket 
          required: false
          schema:
            type: string
        - name: createAt_gte
          in: query
          description: createAt grater
          required: false
          schema:
            type: string
        - name: createAt_lte
          in: query
          description: createAt litter 
          required: false
          schema:
            type: string
        - name: plate
          in: query
          description: plate
          required: false
          schema:
            type: string
        - name: plate_like
          in: query
          description: plate模糊查询
          required: false
          schema:
            type: string
        - name: crm
          in: query
          description: vehicle crm 
          required: false
          schema:
            type: string
        - name: vin
          in: query
          description: vehicle vin 
          required: false
          schema:
            type: string
        - name: createBy
          in: query
          description: 导入者
          required: false
          schema:
            type: string
        - name: brands
          in: query
          description: vehicle brands 支持数组查询
          required: false
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: terminal
          in: query
          description: vehicle terminal 支持数组查询
          required: false
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: park
          in: query
          description: vehicle park 支持数组查询
          required: false
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: status
          in: query
          description: vehicle ticket status 支持数组查询
          required: false
          schema:
            oneOf:
              - type: string
                enum: ["PENDING", "PASSED", "REJECTED"]
              - type: array
                items:
                  type: string
                  enum: ["PENDING", "PASSED", "REJECTED"]
      responses:
        "200":
          description: A paged array of vehicles
          headers:
            X-Total-Count:
              description: Total count of all warnings match the query
              required: true
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /vehicles/{vehicleId}:
    get:
      summary: Find vehicle by id
      operationId: getVehicle
      tags:
        - vehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          description: The id of the vehicle to retrieve
          schema:
            type: string
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: Update vehicle
      operationId: updateVehicle
      tags:
        - vehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          description: The id of vehicle
          schema:
            type: string
      requestBody:
        description: Vehicle to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleDoc"
      responses:
        "200":
          description: The vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      description: deletes a single vehicle based on the ID supplied
      operationId: deleteVehicle
      tags:
        - vehicle
      parameters:
        - name: vehicleId
          in: path
          description: ID of vehicle to delete
          required: true
          schema:
            type: string
      responses:
        "204":
          description: vehicle deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /vehicles/{vehicleId}/alerts:
    get:
      summary: List all alerts by vehicleId
      operationId: listVehicleAlerts
      tags:
        - vehicle
      parameters:
        - name: _limit
          in: query
          description: How many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            maximum: 1000
        - name: _offset
          in: query
          description: How many items to escape
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: _sort
          in: query
          description: Fields to sort
          required: false
          schema:
            type: string
        - name: _select
          in: query
          description: Fields to select
          required: false
          schema:
            type: array
            items:
              type: string
        - name: vehicleId
          in: path
          required: true
          description: The id of vehicle
          schema:
            type: string
        - name: level
          in: query
          description: 根据等级查询报警
          required: false
          schema:
            oneOf:
              - type: array
                items:
                  type: integer
                  format: int32
                  minLength: 1
              - type: integer
                format: int32
                minLength: 1
      responses:
        "200":
          description: A paged array of vehicle alerts
          headers:
            X-Total-Count:
              description: Total count of all warnings match the query
              required: true
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
security:
  - bearerAuth: []
components:
  responses:
...
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    InternalError:
      description: unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
  schemas:
...
    VehicleDoc:
      type: object
      properties:
        ticket:
          type: string
          description: 关联的工单 id
        status:
          type: string
          enum: ["CREATING", "FILLING", "PENDING", "PASSED"]
          description: 车辆录入工单状态
        rejected:
          type: boolean
          description: 工单曾经被驳回
        scrapped:
          type: boolean
          description: 车辆是否报废
        crm:
          type: string
          description: CRM 订单号
        terminal:
          type: string
          description: 终端用户
        park:
          type: string
          description: 车场
        purchaseAt:
          type: string
          format: date-time
          description: 购买日期
        plateAt:
          type: string
          format: date-time
          description: 上牌日期
        operateAt:
          type: string
          format: date-time
          description: 运营日期
        retireAt:
          type: string
          format: date-time
          description: 报废日期
        year:
          type: number
          description: 年份
        license:
          $ref: "#/components/schemas/FileList"
          description: 车辆行驶证
        brands:
          type: string
          description: 整车品牌
        plate:
          type: string
          description: 车牌号
        vin:
          type: string
          description: VIN号
        jobNum:
          type: string
          description: 车辆工号
        num:
          type: string
          description: 车辆自编号
        catl:
          type: string
          description: CATL项目名称
        batteryPN:
          type: string
          description: 电池PN
        batteryNum:
          type: string
          description: 电池号
        batteryPropRight:
          type: string
          description: 电池产权
        power:
          type: number
          description: 功率
        bmuHardware:
          type: string
          description: BMU硬件版本
        bmuSoftware:
          type: string
          description: BMU软件版本
        cscHardware:
          type: string
          description: CSC硬件版本
        cscSoftware:
          type: string
          description: CSC软件版本
        rdb:
          type: string
          description: RDB
        remark:
          type: string
          description: 备注
        softwarePic:
          $ref: "#/components/schemas/FileList"
          description: 上位机软件图片
        frontPic:
          $ref: "#/components/schemas/FileList"
          description: 正面车头图片
        nameplatePic:
          $ref: "#/components/schemas/FileList"
          description: 车辆铭牌图片
        meterPic:
          $ref: "#/components/schemas/FileList"
          description: 车辆仪表图片
        batteryPic:
          $ref: "#/components/schemas/FileList"
          description: 电池号图片
        otherPic:
          $ref: "#/components/schemas/FileList"
          description: 其他图片
        createBy:
          type: string
          description: 录入者
    VehicleCreateDoc:
      type: object
      required:
        - terminal
        - purchaseAt
        - plateAt
        - operateAt
        - year
        - brands
        - plate
        - vin
        - jobNum
        - catl
        - batteryPN
        - batteryNum
        - power
        - bmuHardware
        - bmuSoftware
        - cscHardware
        - cscSoftware
        - rdb
        - softwarePic
        - frontPic
        - nameplatePic
        - meterPic
        - batteryPic
        - createBy
      properties:
        terminal:
          type: string
        park:
          type: string
        purchaseAt:
          type: string
          format: date-time
        plateAt:
          type: string
          format: date-time
        operateAt:
          type: string
          format: date-time
        retireAt:
          type: string
          format: date-time
        year:
          type: number
        license:
          $ref: "#/components/schemas/FileList"
        brands:
          type: string
        plate:
          type: string
        vin:
          type: string
        jobNum:
          type: string
        num:
          type: string
        catl:
          type: string
        batteryPN:
          type: string
        batteryNum:
          type: string
        batteryPropRight:
          type: string
        power:
          type: number
        bmuHardware:
          type: string
        bmuSoftware:
          type: string
        cscHardware:
          type: string
        cscSoftware:
          type: string
        rdb:
          type: string
        remark:
          type: string
        status:
          type: string        
        softwarePic:
          $ref: "#/components/schemas/FileList"
        frontPic:
          $ref: "#/components/schemas/FileList"
        nameplatePic:
          $ref: "#/components/schemas/FileList"
        meterPic:
          $ref: "#/components/schemas/FileList"
        batteryPic:
          $ref: "#/components/schemas/FileList"
        otherPic:
          $ref: "#/components/schemas/FileList"
        createBy:
          type: string    
    Vehicle:
      allOf:
        - $ref: "#/components/schemas/VehicleDoc"
        - $ref: "#/components/schemas/MongoDefault"
    FileDoc:
      type: object
      properties:
        name:
          type: string
          description: 原始文件名
        ossName:
          type: string
          description: oss上的文件名
        remark:
          type: string
          description: 备注
        size:
          type: number
          description: 文件大小
        status:
          type: string
          description: 上传状态 
        uid:
          type: string
          description: uid
        url:
          type: string
          description: url
    FileList:
      type: array
      items:
        $ref: "#/components/schemas/FileDoc"
    AlertDoc:
      type: object
      properties:
        level:
          type: integer
          format: int32
          description: "故障等级"
        code:
          type: string
          description: "故障码"
        name:
          type: string
          description: "故障名称"
        startedAt:
          type: string
          format: date-time
          description: "开始报警的时间"
        count:
          type: integer
          format: int32
          description: "次数"
    Alert:
      allOf:
        - $ref: "#/components/schemas/AlertDoc"
        - $ref: "#/components/schemas/MongoDefault"
...
    MongoDefault:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        updateAt:
          type: string
          format: date-time
        updateBy:
          type: string
        createAt:
          type: string
          format: date-time
        createBy:
          type: string
    Err:
      type: object
      required:
        - name
        - message
      properties:
        code:
          type: string
        type:
          type: string
        message:
          type: boolean
        name:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              keyword:
                type: string
              message:
                type: string
              path:
                type: string
              value:
                type: string
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT